Default values to configuration properties ("LoopConfig").
Adding default values to several configuration parameters
according to their description in schema.
Thus they will be initialized at the time of deserializing
(creation), rather than handling this in the code.
This simplifies the code, since the relevant "Option"s' values
will never be "None", and also makes the code more error prone.
1. The "delay" sub-property was not really optional until now.
Fixed it to be optional with a default value of 0.
2. The whole "struct LoopConfig" has now default values.
This saves extra (& duplicate) code to check if it was omitted in the
".yaml" configuration file and initialize to 0 "delay" and 1 "times".

Signed-off-by: Pavel Bar <pbar@redhat.com>