Enhanced environment variables' support (commands).
Added support to use environment variables in the commands.
Motivation / use-case:
You can run commands/scripts, also remotely by providing them
parameters that are received via environment variables declared on
the machine where you run the "autopilot" tool.
This is useful when you have a script that receives an input argument
(i.e., IP) and you don't want to manually update the script or the
already written "autopilot" flow "yaml" file.
Example: run a script at the ASIL layer that receives the IP of
your laptop running Grafana and configures the collector service
which is running at QM layer to be able to propagate the required data
all the way to be displayed at Grafana dashboards.
------
...
command: config_collector.sh $env:GRAFANA_IP
...
------
As part of the effort, a minor refactoring was performed to
store the command inside the "CommandSession" object.
Now the latter is also responsible to resolve all the environment
variables (if there are any).

Signed-off-by: Pavel Bar <pbar@redhat.com>